Match-id-1568f48f7821682b613a6a20e0457f53cc05a93a

+++ mef-center-install/pkg/control/exchange_ca.go
@@ -149,16 +149,16 @@ func (ecf *ExchangeCaFlow) importCa() error {
 	if err := ecf.copyCaToCertManager(); err != nil {
 		return err
 	}
-
-	// delete old crl and backup of old crl
-	crl := ecf.pathMgr.ConfigPathMgr.GetNorthernCrlPath()
-	crlBackup := crl + backuputils.BackupSuffix
-	for _, filePath := range []string{crl, crlBackup} {
-		if err := fileutils.DeleteFile(filePath); err != nil {
-			return fmt.Errorf("clear old crl failed, error: %s", err.Error())
+	if ecf.component == util.NginxManagerName {
+		// delete old crl and backup of old crl
+		crl := ecf.pathMgr.ConfigPathMgr.GetNorthernCrlPath()
+		crlBackup := crl + backuputils.BackupSuffix
+		for _, filePath := range []string{crl, crlBackup} {
+			if err := fileutils.DeleteFile(filePath); err != nil {
+				return fmt.Errorf("clear old crl failed, error: %s", err.Error())
+			}
 		}
 	}
-
 	hwlog.RunLog.Infof("import [%s] cert success", ecf.component)
 	return nil
 }
